Data Preparation for Fixed Effects Modelling

Data Preparation for Fixed Effects Modelling

fe_data_prep(Y, Z, ID, cutoff = 10, check = TRUE)

Arguments

Y

a numerical vector, with values of 0 or 1, indicating the outcome variable.

Z

a matrix or data frame containing covariates.

ID

a vector representing the provider id. Its elements can be either numeric values or characters.

cutoff

an integer as cutoff of provider size with 10 as default. Facilities with observations fewer than the "cutoff" value will be labeled as "include = 0".

check

a Boolean indicating whether checking missingness, variation, correlation and VIF of variables in the data. Defaulting to "TRUE".

...

Value

data

a data frame including response, provider ID, and covariates

char_list

a list including variable names.

Details

Major steps in this stage include (in order): * checking missingness, variation, correlation and VIF of variables in the data, * provider screening based on number of discharges, 10 as default, * calculating the total number of facilities and hospitals, * reporting proportions of facilities with no readmission within 30 days, and with all readmission within 30 days after discharge, * modifying responses to prevent infinite log-likelihoods during model fitting, and * sorting data by hospital and provider identifiers.

The `fe_data_prep()` function not only returns the original data sorted by provider identifiers, but also outputs vectors that indicate whether the provider has zero or all readmissions, and a dataframe combines all the output information is also provided.

Examples

data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID)
#> Checking missingness of variables ... 
#> Missing values NOT found. Checking missingness of variables completed!
#> Checking variation in covariates ... 
#> Checking variation in covariates completed!
#> Checking pairwise correlation among covariates ... 
#> Checking pairwise correlation among covariates completed!
#> Checking VIF of covariates ... 
#> Checking VIF of covariates completed!
#> Warning: 0 out of 10 providers considered small and filtered out!
#> 0 out of 10 remaining providers with no readmission within 30 days.
#> 0 out of 10 remaining providers with all readmissions within 30 days.
#> After screening, 52% of all discharges were readmitted within 30 days.
head(data.prep$data)
#>    Y ID      Z1      Z2      Z3      Z4 included no.readm all.readm
#> 38 1  1 -0.7124  0.3953 -0.8402  0.0840        1        0         0
#> 6  0  1  0.3660 -0.7654  1.8015 -0.5231        1        0         0
#> 2  0  1 -0.0715  0.3307 -0.1121 -0.3115        1        0         0
#> 32 1  1  1.2804 -0.7822  0.0941  0.7869        1        0         0
#> 34 0  1  0.4012 -0.5114  0.3514 -0.8059        1        0         0
#> 23 0  1  0.8727 -0.7320 -0.2893 -0.0106        1        0         0
data.prep$char_list
#> $Y.char
#> [1] "Y"
#> 
#> $prov.char
#> [1] "ID"
#> 
#> $Z.char
#> [1] "Z1" "Z2" "Z3" "Z4"
#>