Make skeleton package to test roclet system

Make a dummy package skeleton including 'srr' roxygen2 tags which can be used to try out the functionality of this package. Running the example lines below which activate the 'srr' roclets, and show you what the output of those roclets looks like. Feel free to examine the effect of modifying any of the @srrstats tags within the code as identified by running those lines.

srr_stats_pkg_skeleton(base_dir = tempdir(), pkg_name = "demo")

Arguments

base_dir
The base directory where the package should be constructed.

pkg_name
The name of the package. The final location of this package will be in file.path(base_dir, pkg_name).

Value

The path to the directory holding the newly created package

See also

Examples

d <- srr_stats_pkg_skeleton (pkg_name = "mystatspkg")
# (capture.output of initial compliation messages)
x <- capture.output (roxygen2::roxygenise (d), type = "output")
#> ℹ Loading mystatspkg
#> Re-compiling mystatspkg
#> ──────────────────── rOpenSci Statistical Software Standards ───────────────────
#> 
#> ── @srrstats standards: 
#>   * [G1.1, G1.2, G1.3, G2.0, G2.1] in function 'test_fn()' on line#11 of file [R/test.R]
#>   * [RE2.2] on line#2 of file [tests/testthat/test-a.R]
#>   * [G2.3] in function 'test()' on line#6 of file [src/cpptest.cpp]
#>   * [G1.4] on line#17 of file [./README.Rmd]
#> 
#> ── @srrstatsNA standards: 
#>   * [RE3.3] on line#5 of file [R/srr-stats-standards.R]
#> 
#> ── @srrstatsTODO standards: 
#>   * [RE4.4] on line#14 of file [R/srr-stats-standards.R]
#>   * [RE1.1] on line#11 of file [R/test.R]
#>   * [G1.5] on line#17 of file [./README.Rmd]
#> ────────────────────────────────────────────────────────────────────────────────